Portals4 runtime builds without errors.
runtime/libupc/portals4
	gupcr_backtrace.c
	gupcr_backtrace.h
	gupcr_nb.upc
	gupcr_nb_sup.c
	gupcr_nb_sup.h
	- Added non-blacking and backtrace support

	gupcr_coll_init.upc
	gupcr_coll_reduce.in
	gupcr_coll_reduce.upc
	gupcr_config.h
	gupcr_gmem.c
	gupcr_pts.h
	upc-lib.in
	- Various fixes to compile with clang

runtime/libupc
	CMakeLists.txt
	config.h.cmake
	- Added missing configuration variables
	- Added build of reduce collective tests

+++ runtime/libupc/portals4/gupcr_gmem.c
@@ -181,7 +181,7 @@ void
 gupcr_gmem_get (void *dest, int thread, size_t offset, size_t n)
 {
   ptl_process_t rpid;
-  char *dest_addr = (char *) (dest - USER_PROG_MEM_START);
+  char *dest_addr = (char *)dest - (size_t) USER_PROG_MEM_START;
   size_t n_rem = n;
 
   gupcr_debug (FC_MEM, "%d:0x%lx 0x%lx",
